requirements.txt
PyPDF2>=3.0.0
python-docx>=0.8.11
edge-tts>=6.1.9
moviepy>=1.0.3
